refactor(backend): isolate analytical answer prompt
- Move the late analytical answer contract into a focused prompt helper
- Preserve the direct_prompts compatibility import surface for graph/runtime bindings
- Update the runtime cleanup audit for issue #547
